move linter to file
